Show only the first part of the story on the page.
I apply vertical margins at different viewport sizes so that only the first part of the storyline shows when the page loads. However, the content is not vertically centered on the page, which is bothersome. How do I vertically align the content in the center of the page, but also only show one part of the story at a time? Maybe I need a whole different approach. How do I figure out what approach to take? I have seen other sites do this, but I am not sure how it is done. How could I figure it out?